Add files via upload
In the script there are two files  final_ensemble_report_all_clean.csv and protein_synonyms.csv, where final_ensemble_report_all_clean.csv has the relations that we found from model and protein_synonyms.csv contains each protein’s UniProt ID, gene names, and synonyms. The script builds a lookup table to map every protein to all possible names that might identify it in the STRING database. So you can see Used_Name_1 and Used_Name_2 column in the output csv file which indicates the names that were found to get the relations using API. Once both proteins are successfully matched to STRING IDs, it asks string’s interaction network endpoint to check whether string reports an interaction between the two proteins or not.If an interaction exists, the script marks "STRING_Found" = True and records the string confidence score. Otherwise it stores False and scores 0. The script does not use Ensemble_Result to define the interaction because string does not records the relation types, rather it only records the proteins has interections or not. That’s why if the exact relation types are tested, it always returns false. So instead, Ensemble_Result is just copied to the final output so we can compare our detected relation with STRING’s. Finally, it writes all results.